+++ Packages/com.9b9387.owlet/CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2022-11-08
+## [0.0.0] - 2022-11-08
 
 ### This is the first release of *\<Owlet\>*.
 
-*Short description of this release*
+- Wrapper Asset Loader
+- Wrapper UI Manager
+- Wrapper WebRequest Manager
+- Wrapper Event Center